+++ xarray/core/dataset.py
@@ -4956,11 +4956,10 @@ class Dataset(
         self,
         coord_names: str | Sequence[Hashable],
         index_cls: type[Index] | None = None,
-        drop_existing: bool = False,
         **options,
     ) -> Self:
         """Set a new, Xarray-compatible index from one or more existing
-        coordinate(s).
+        coordinate(s). Existing index(es) on the coord(s) will be replaced.
 
         Parameters
         ----------
@@ -4971,9 +4970,6 @@ class Dataset(
             The type of index to create. By default, try setting
             a ``PandasIndex`` if ``len(coord_names) == 1``,
             otherwise a ``PandasMultiIndex``.
-        drop_existing : bool
-            Whether to drop indexes on any existing coord_names if one
-            is present
         **options
             Options passed to the index constructor.
 
@@ -5008,18 +5004,6 @@ class Dataset(
                     f"those variables are data variables: {data_vars}, use `set_coords` first"
                 )
             raise ValueError("\n".join(msg))
-
-        # we could be more clever here (e.g., drop-in index replacement if index
-        # coordinates do not conflict), but let's not allow this for now
-        indexed_coords = set(coord_names) & set(self._indexes)
-
-        if indexed_coords:
-            if drop_existing:
-                self.drop_indexes(indexed_coords)
-            else:
-                raise ValueError(
-                    f"those coordinates already have an index: {indexed_coords}"
-                )
 
         coord_vars = {name: self._variables[name] for name in coord_names}