+++ doc/whats-new.rst
@@ -14,9 +14,9 @@ v2025.12.1 (unreleased)
 New Features
 ~~~~~~~~~~~~
 
-- Added ``drop_existing`` parameter to :py:meth:`Dataset.set_xindex` and
-  :py:meth:`DataArray.set_xindex` to allow replacing existing indexes without
-  needing to call :py:meth:`drop_indexes` first (:pull:`11008`).
+- :py:meth:`Dataset.set_xindex` and :py:meth:`DataArray.set_xindex`
+  automatically replace any existing index being set instead of erroring
+  or needing needing to call :py:meth:`drop_indexes` first (:pull:`11008`).
   By `Ian Hunt-Isaak <https://github.com/ianhi>`_.
 
 Breaking Changes